+++ R/estimate-functions.R
@@ -85,17 +85,17 @@ estimate_k_1RM <- function(weight, reps, eRIR = 0, reverse = FALSE, weighted = "
 
   if (reverse == FALSE) {
     m1 <- stats::nls(
-      nRM ~ ((1 / k) / (weight / `0RM`)) - (1 / k),
+      nRM ~ (k * `1RM` + `1RM` - weight)/(k  * weight),
       data =  df,
-      start = list(k = 1, `0RM` = max(df$weight)),
+      start = list(k = 1, `1RM` = max(df$weight)),
       weights = df$reg_weights,
       ...
     )
   } else {
     m1 <- stats::nls(
-      weight ~ `0RM` / (k * nRM + 1),
+      weight ~ ((k + 1) * `1RM`)/(k * nRM + 1),
       data =  df,
-      start = list(k = 1, `0RM` = max(df$weight)),
+      start = list(k = 1, `1RM` = max(df$weight)),
       weights = df$reg_weights,
       ...
     )
@@ -278,37 +278,4 @@ estimate_klin_1RM <- function(weight, reps, eRIR = 0, reverse = FALSE, weighted 
   }
 
   m1
-}
-
-#' @describeIn estimate_functions Estimate the 1RM from \code{\link{estimate_k_1RM}} function
-#'
-#' The problem with Epley's estimation model (implemented in \code{\link{estimate_k_1RM}} function)
-#'     is that it predicts the 1RM when nRM = 0. Thus, the estimated parameter in the model produced
-#'     by the \code{\link{estimate_k_1RM}} function is not 1RM, but 0RM. This function calculates the
-#'     weight at nRM = 1 for both the normal and reverse model. See Examples for code
-#'
-#' @param model Object returned from the  \code{\link{estimate_k_1RM}} function
-#' @export
-#' @examples
-#' # ---------------------------------------------------------
-#' # Estimating 1RM from Epley's model
-#' m1 <- estimate_k_1RM(150 * c(0.9, 0.8, 0.7), c(3, 6, 12))
-#' m2 <- estimate_k_1RM(150 * c(0.9, 0.8, 0.7), c(3, 6, 12), reverse = TRUE)
-#'
-#' # Estimated 0RM values from both model
-#' c(coef(m1)[[1]], coef(m2)[[1]])
-#'
-#' # But these are not 1RMs!!!
-#' # Using the "reverse" model, where nRM is the predictor (in this case m2)
-#' # makes it easier to predict 1RM
-#' predict(m2, newdata = data.frame(nRM = 1))
-#'
-#' # But for the normal model it involve reversing the formula
-#' # To spare you from the math pain, use this
-#' get_predicted_1RM_from_k_model(m1)
-#'
-#' # It also works for the "reverse" model
-#' get_predicted_1RM_from_k_model(m2)
-get_predicted_1RM_from_k_model <- function(model) {
-  stats::coef(model)[[2]] / (stats::coef(model)[[1]] + 1)
 }